pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-coscientist"
version = "0.1.0"
description = "Open AI Co-Scientist framework reimplemented with LangGraph for clean, modular hypothesis generation. Copyright Jataware Corp, 2025."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Brandon Rose", email = "brandon@jataware.com"}
]
license = {file = "LICENSE"}
keywords = ["ai", "research", "hypothesis", "langgraph", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langgraph~=1.0.6",
    "langchain-core~=1.2.7",
    "langchain-mcp-adapters>=0.2.1",
    "langsmith~=0.6.2",
    "litellm~=1.80.16",
    "typing-extensions~=4.15.0",
    "rich~=14.2.0",
    "jsonschema~=4.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest~=9.0.2",
    "pytest-asyncio~=1.3.0",
    "black~=25.12.0",
    "ruff~=0.14.11",
    "mypy~=1.19.1",
    "python-dotenv~=1.2.1",
]

[project.urls]
Homepage = "https://github.com/jataware/open-coscientist"
Documentation = "https://github.com/jataware/open-coscientist#readme"
Repository = "https://github.com/jataware/open-coscientist"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["open_coscientist*"]

[tool.setuptools.package-data]
open_coscientist = ["prompts/*.md"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]